refactor move run function to inference.py; delete unused functions
